made change in services and navbar
