+++ src/Services/AuthInterface.tsx
@@ -1,0 +1,84 @@
+
+export interface IBatchResponse {
+    IbatchArray: IbatchObject[]
+  }
+  
+  export interface IBatchObject {
+    id: number;
+    batch_name: string;
+    student: string[];
+  }
+  
+  export interface ISectionResponse {
+    IsectionArray: ISectionObject[]
+  }
+  
+  export interface ISectionObject {
+    id: number;
+    batch_name: string;
+    student: string[];
+  }
+  
+  
+  export interface IForgotPassword {
+    email: string;
+  }
+
+  export interface IAuthlogin {
+    username: string;
+    password: string;
+    rememberMe: boolean;
+  }
+  
+  export interface IAuthloginResponse {
+    token: string;
+    error: string;
+    status: number;
+    user: {
+      id: number;
+      name: string;
+      email: string;
+      roles: Array<IRoles>
+    };
+  }
+  
+  
+  interface IRoles{
+    id:number,
+    name:string
+    }
+  export interface IAuthsignupResponse {
+    id: number,
+    name: string,
+    email: string,
+    roles:Array<IRoles>
+  }
+  export interface IStudentAccountCreate {
+    name: string;
+    batch: string;
+    section: string;
+    email: string;
+    password: string;
+  }
+  export interface IAdminAccountCreate {
+    name: string;
+    email: string;
+    password: string;
+  }
+  export interface IBatchResponse {
+    IbatchArray: IbatchObject[];
+  }
+  export interface IbatchObject {
+    id: number;
+    batch_name: string;
+    student: string[];
+  }
+  export interface ISectionResponse {
+    IsectionArray: ISectionObject[];
+  }
+  export interface ISectionObject {
+    id: number;
+    batch_name: string;
+    student: string[];
+  }
+
